requirements.txt
Pillow
pyscreenshot
pyqt5